Add Storybook tests (#137)
Prevent regressions in our custom components by adding Storybook
testing. Copy Storybook tests from the archived `gravitational/docs`
repo. Add a rough-and-ready Storybook configuration to enable tests to
pass.

One significant complication is that Docusaurus generates a Webpack
configuration when building a docs site. There are Storybook frameworks
and add-ons for Docusaurus that take advantage of Docusaurus's
asset-loading logic, but none are currently being maintained. The
quickest thing we can do is add a separate Webpack configuration that
renders components in Storybook similarly (but not identically) to the
Docusaurus site.

A separate change can refine this approach by, for example:
- Vendoring a Storybook Docusaurus framework
- Migrating Storybook tests to `react-testing-library`

Also add a step to the `Run tests` GitHub Actions job to run the
Storybook tests.