Remove remark-update-tags (#171)
Closes #91
Closes #92

After we removed most of `remark-update-tags`, the remaining logic in
the plugin updates legacy `TabItem` tags to use the Docusaurus `TabItem`
component by adding a unique `value` prop to each `TabItem`.

This change removes `remark-update-tags` by moving the `value`-adding
logic to the `Tabs` component. It uses a wrapper to assign a unique
`value` to each `TabItem`. This way, we can keep the tab alteration
logic where a `docs-website` contributor might expect it, close to the
component code, rather than in a plugin that alters an AST.

Note that the Docusaurus `Tabs` component checks each `TabItem` for the
`value` prop, so it is not possible to assign the `value` within
`TabItem`.