Fix the markdown-lint test (#178)
In the `Run tests` GitHub Actions job, we run the `yarn markdown-lint`
script to ensure that changes to custom remark plugins haven't caused
the remark config, `.remarkrc.mjs`, to break. Since we run the remark
linter in the `gravitational/teleport` CI, this is important to prevent
that pipeline from breaking.

In the current `Run tests` config, we run `yarn markdown-lint` without
loading submodule content. This is problematic since `remark` only
executes the JavaScript in `remarkrc.mjs` if there is at least one file
to lint. This means that PRs that introduced runtime errors into the
remark config, such as #158, manage to pass the linter, since there were
no files to lint in the `Run tests` job.

This change adds a script that writes mock files to a docs content
directory before running `yarn markdown-lint`. This way, we can test the
`.remarkrc.mjs` config without performing a full submodule load, and the
test won't depend on whether a content change happens to introduce a
linter error.