Add a link for creating a GitHub issue (#205)
* Add a link for creating a GitHub issue

Closes #110

In the previous docs site (github.com/gravitational/docs), a "Create
issue on GitHub button" was a valuable source of feedback. This change
adds a link that opens a new tab with the GitHub Create Issue form,
prepopulated with the URL path of the current docs page.

The best positioning for such a link is not obvious. In the previous
docs site, we placed the button on the right sidebar under the table of
contents for a given docs page. However, on the Docusaurus site, in-page
tables of contents sometimes span the entire viewport height, e.g., in
the main FAQ page with its 20 H2 headers, making a button on the right
sidebar difficult to find.

Further, Docusaurus makes some styling assumptions about the right
sidebar, like the fact that the `TOCItems` component includes a left
border but the column containing it does not, that make it non-trivial
to style a new component there. We would need to rework the styling of
the entire right sidebar component layout.

The left sidebar is also not an ideal place to add this link, since it
handles navigation for the entire docs site, and it would be unexpected
to find a component there that is scoped to a single page.

The solution this change proposes is to add the GitHub issue link below
the main title of each page, swizzling the `@theme/DocItem/Content`
component. This approach requires minimal reworking of the Docusaurus
component layout, but still offers users a highly visible option to
report an issue that, if a user is not interested in reporting an issue,
remains relatively unobtrusive.

This change adds a self-contained `GitHubIssueLink` component, and we
can move this to a more suitable location (or turn it from a link into a
button) when time opens up.

* Use right alignment for the link

Respond to roraback feedback.

* Apply roraback feedback

Left-align the link and include a smaller marging with the H1.