Validate redirects instead of rewriting them (#219)
In order to migrate the legacy custom site to Docusaurus, we added logic
to the Docusaurus config to convert redirects to use the Docusaurus
format. The logic included some steps to rewrite redirect sources and
desinations.

This change adjusts the logic to validate redirects instead and throw
errors on malformed redirects. This way, it is easier to troubleshoot
issues with redirects, since the redirects that a documentation author
would see are the same ones that Docusaurus processes.

As part of this, this change adds a rule to require a leading slash in
redirect paths. When validating redirects, Docusaurus ignores
destinations that do not begin with a slash, causing unexpected issues
for custom logic in the docs engine that expects redirect paths to begin
with slashes.

For context, see the following file in facebook/docusaurus#9171:

packages/docusaurus-plugin-client-redirects/src/collectRedirects.ts