Fix details rendering in remark-includes (#246)
Closes #238

We use the `remark-includes` plugin to render partials in the docs. To
render a partial, the plugin operates on a Markdown AST, loading the
partial, parsing it into a sub-tree, then inserting it into the parent
AST. While parsing a partial that includes a `<details>` element, the
plugin inserts the details box's `<summary>` element into a `<p>`
element.

As a result, `<details>` boxes included in partials do not render their
`<summary>` elements as expected in Docusaurus. The Docusaurus `Details`
component uses the value of the component's first-level `<summary>`
child element if it is present and, if not, uses the default value
"Details" for the summary. Since `remark-includes` wraps the `<summary>`
tag in a `<p>` tag, the `<summary>` is no longer a first-level child of
the `Details` component.

This change edits `remark-includes` to, after parsing a partial, find
any `<details>` elements in the partial and elevate their summaries to
the top level of each details box.

This change also restores tests to `remark-includes`. While we had these
in the original github.com/gravitational/docs repository, we removed
them when migrating to Docusaurus. Since the post-migration version of
the plugin has a different function signature than the original,
including a callback function to update relative paths in partials, this
change adds a test version of this function. This change also adds new
tests for rendering `<details>` boxes.