Emit gtag events on code copy button clicks (#260)
* Emit gtag events on code copy button clicks

Help understand how documentation readers use docs pages to perform
operations on a Teleport cluster by adding an anonymous gtag event when
a user clicks a code snippet copy button. This includes both the large
button that copies an entire multi-line snippet and the smaller button
that appears next to a single command within a `code` snippet.

To begin with, gtag event payloads are minimal and contain no
information about the content of the snippet:
- `label`: the language of the code snippet, e.g., `code` or `yaml`
- `scope`: whether the copy button is for a single line or the whole
  snippet
- Default fields such as the path of the page.

As part of this work, include an optional `gtag` prop in components that
emit a gtag so we can easily and explicitly assert against gtag events
in tests.

* Refactor gtag calls to use `trackEvent`

Use the `trackEvent` function added in #250 to send gtag events. Make sure we
are using a single function to send these events in order to keep our
monitoring consistent. This change adds to `trackEvent` in order to
allow for mock `gtag` functions in testing. If there is no `gtag`
function available, `trackEvent` logs the event to the console.

Also move code from `src/utils/gtag.ts` to `src/utils/analytics.ts` so
all analytics-related code is in one place.

* Record index and count in code copy button events

Indicate how far down a page a `Command` line or `Pre` snippet is when a
user clicks a code copy button, including both the whole-snippet button
and the individual line button.

Add a `PositionProvider` context provider. Components register
themselves with the provider when rendering so the parent component
doesn't need to know about its children in advance. This is necessary
for docs pages, since a `DocItem` doesn't know in advance of MDX parsing
how many `Pre`s are in the page. Similarly, the `Snippet` component
results from the `remark-hljs` plugin and doesn't know how many
`Command`s are in a code snippet.

Keep the total count of elements the same between renders by using
`useRef` and `nanoid` to give each `Pre` and `Command` a unique ID that,
unlike the result of `useID`, stays the same between renders.

To track the position of each `Command` within a `Pre`, have `Pre` wrap
its `Command`s in a nested `PositionProvider`. Pass a callback to
`PositionProvider` so it can report the total number of outer `Pre`s, as
well as the position of each `Pre`.