Use best practices for TabItem values (#321)
The change to `Tabs` rendering in #234 used `useId` to assign a `value`
to each `TabItem` within a `Tabs`. This goes against the React best
practice of calling hooks at the top level of a component, and could
lead to unexpected behavior if the number of children in a `Tabs`
becomes dynamic. This change uses the index of each `TabItem` within a
`Tabs` for the `value` instead.