Add tags.yml enforcement (#349)
Now that the docs site is taking advantage of the Docusaurus tagging
feature, it makes sense to restrict the tags that docs authors can add
to docs pages. This way, we can ensure that a user who navigates to the
article list for a specific tag can see a substantial cross-section of
the docs site.

Docusaurus allows sites to include a `tags.yml` file, a list of allowed
tags, at the root of each content directory. The `tags.yml` file also
provides descriptions of each tag. The docs engine includes these as the
first paragraphs of each tag page.

- Add a `tags.yml` file with allowed tags for guide type, product, cloud
  provider, and use case
- In `scripts/prepare-files.mts`, copy `tags.yml` to the `docs`
  directory and each directory within `versioned_docs`