Add a remark linter to require intro paragraphs (#383)
Introductory paragraphs are critical to laying out the scope of a docs
page and helping a user determine if they should keep reading. However,
docs pages sometimes do not include introductory paragraphs. This change
adds a remark linter to require introductory paragraphs before the first
top-level heading of a docs page.

Note that while `gravitational/teleport` currently includes a Vale
linter for first paragraphs, it is not possible for authors to ignore it
using comments because it has the `raw` scope. The linter added by this
change supports ignoring via comments. We can also fine-tune the
behavior more precisely since it acts on a Markdown AST, rather than
text.