Add a remark linter for improper `Var` use (#387)
See #204

We currently use a Vale linter to catch improper use of the `Var`
component, i.e., pages in which there is a single instance of a `Var`
with a particular name. This linter catches `Var`s that are:
- Difficult to find and use, or potentially unnecessary, since there are
  no instructions to assign the variable, and no repeated values to
  assign simultaneously with a `Var` in the docs page.
- Mistakenly inconsistent, with names that the author intended to be
  identical but actually differ.

This change adds a `remark` linter to achive this. The `remark` linter
has several advantages over the current Vale linter, including:
- It shows the name of the `Var` that violated the linter, making it
  much easier to use.
- It can be ignored with comments. The Vale linter cannot be ignored
  with comments since it uses the `raw` scope, and only Vale linters
  without the `raw` scope can be ignored with comments.
- We can use unit tests to verify that it works and catch edge cases.