Allow tab item synchronization (#436)
In this change, when a reader selects a tab item with a given label, the
docs site selects all tab items in the docs with the same label. This
makes it easier for a reader to interact with docs pages with a lot of
tabs, since they do not need to continually watch for incorrect tab
selections and click the correct tab item on each `Tabs` component they
encounter.

Docusaurus has built-in tab synchronization. `Tabs` components with the
same `groupId` prop render tab items with the same `value` as selected
[1]. The wrapped `Tabs` component we include in the docs site source
already assigns `Tabs` props automatically. This change adjusts the prop
assignment logic to assign every `Tabs` component the same `groupId`. It
also assigns every tab item a `value` based on its label. This lets us
synchronize every tab selection in the docs, even when the tab items in
one `Tabs` include labels that other `Tabs` don't, without requiring any
new work on the part of docs authors.

[1]:
https://docusaurus.io/docs/markdown-features/tabs#syncing-tab-choices